+++ src/main/java/tn/esprit/projectclass/services/IGeneric.java
@@ -4,5 +4,5 @@ import java.util.List;
 public interface IGeneric<T>{
     List<T> findAll() throws Exception;
     T save(T entity) throws Exception;
-    void delete(Long id) throws Exception;
+    void delete(int id) throws Exception;
 }
